refactor: extract registration JSON parsing into shared utility
